refactor(web): extract shared useOptimisticPatchMutation helper
useUpdateNotificationPreferences and useUpdateLobbyNotificationPreferences
each duplicated the same onMutate-snapshot/setQueryData-patch/onError-
rollback pattern for merging a partial update into a single cached
object. Extracted useOptimisticPatchMutation() and had both use it.
