+++ lined-web/src/hooks/useNotifications.ts
@@ -1,4 +1,4 @@
-import { useQuery, useMutation, useQueryClient } from '@tanstack/react-query';
+import { useQuery } from '@tanstack/react-query';
 import {
   getPreferences,
   updatePreferences,
@@ -6,6 +6,7 @@ import {
   updateLobbyPreferences,
 } from '@/api/notifications';
 import { QUERY_KEYS } from '@/lib/constants';
+import { useOptimisticPatchMutation } from './useOptimisticPatchMutation';
 import type {
   NotificationPreferencesDto,
   NotificationPreferencesUpdateDto,
@@ -19,45 +20,11 @@ export const useNotificationPreferences = () =>
     queryFn: getPreferences,
   });
 
-interface UpdatePreferencesContext {
-  previous: NotificationPreferencesDto | undefined;
-}
-
-export const useUpdateNotificationPreferences = () => {
-  const queryClient = useQueryClient();
-  return useMutation<
-    NotificationPreferencesDto,
-    unknown,
-    NotificationPreferencesUpdateDto,
-    UpdatePreferencesContext
-  >({
+export const useUpdateNotificationPreferences = () =>
+  useOptimisticPatchMutation<NotificationPreferencesDto, NotificationPreferencesUpdateDto>({
+    queryKey: QUERY_KEYS.notificationPreferences,
     mutationFn: updatePreferences,
-    onMutate: async (patch) => {
-      await queryClient.cancelQueries({ queryKey: QUERY_KEYS.notificationPreferences });
-      const previous = queryClient.getQueryData<NotificationPreferencesDto>(
-        QUERY_KEYS.notificationPreferences,
-      );
-      if (previous) {
-        queryClient.setQueryData<NotificationPreferencesDto>(
-          QUERY_KEYS.notificationPreferences,
-          { ...previous, ...patch },
-        );
-      }
-      return { previous };
-    },
-    onSuccess: (updated) => {
-      queryClient.setQueryData<NotificationPreferencesDto>(
-        QUERY_KEYS.notificationPreferences,
-        updated,
-      );
-    },
-    onError: (_err, _patch, context) => {
-      if (context?.previous) {
-        queryClient.setQueryData(QUERY_KEYS.notificationPreferences, context.previous);
-      }
-    },
   });
-};
 
 export const useLobbyNotificationPreferences = (lobbyId: number) =>
   useQuery({
@@ -65,38 +32,11 @@ export const useLobbyNotificationPreferences = (lobbyId: number) =>
     queryFn: () => getLobbyPreferences(lobbyId),
   });
 
-interface UpdateLobbyPreferencesContext {
-  previous: LobbyNotificationPreferencesDto | undefined;
-}
-
-export const useUpdateLobbyNotificationPreferences = (lobbyId: number) => {
-  const queryClient = useQueryClient();
-  const queryKey = QUERY_KEYS.lobbyNotificationPreferences(lobbyId);
-  return useMutation<
+export const useUpdateLobbyNotificationPreferences = (lobbyId: number) =>
+  useOptimisticPatchMutation<
     LobbyNotificationPreferencesDto,
-    unknown,
-    LobbyNotificationPreferencesUpdateDto,
-    UpdateLobbyPreferencesContext
+    LobbyNotificationPreferencesUpdateDto
   >({
+    queryKey: QUERY_KEYS.lobbyNotificationPreferences(lobbyId),
     mutationFn: (data) => updateLobbyPreferences(lobbyId, data),
-    onMutate: async (patch) => {
-      await queryClient.cancelQueries({ queryKey });
-      const previous = queryClient.getQueryData<LobbyNotificationPreferencesDto>(queryKey);
-      if (previous) {
-        queryClient.setQueryData<LobbyNotificationPreferencesDto>(queryKey, {
-          ...previous,
-          ...patch,
-        });
-      }
-      return { previous };
-    },
-    onSuccess: (updated) => {
-      queryClient.setQueryData<LobbyNotificationPreferencesDto>(queryKey, updated);
-    },
-    onError: (_err, _patch, context) => {
-      if (context?.previous) {
-        queryClient.setQueryData(queryKey, context.previous);
-      }
-    },
   });
-};
